Gets JNI Working
