+++ src/main/groovy/edu/wpi/first/nativeutils/ExportsConfig.groovy
@@ -10,4 +10,7 @@ interface ExportsConfig extends Named {
 
     void setX64ExcludeSymbols(List<String> symbols)
     List<String> getX64ExcludeSymbols();
+
+    void setExcludeBuildTypes(List<String> excludes)
+    List<String> getExcludeBuildTypes()
 }
